move the logic of photo processing to processImages.ts
